requirements.txt
anyio==3.6.2
attrs==22.2.0
Automat==22.10.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.95.1
filelock==3.9.0
h11==0.14.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==23.0
parsel==1.7.0
passlib==1.7.4
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyDispatcher==2.0.7
pymongo==4.3.3
pyOpenSSL==23.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rsa==4.9
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.19.0
websockets==11.0.2
zope.interface==5.5.2
